Support a custom username/password for basic authentication
When using Auth0 authentication for Jenkins we don't have a way to give
the scraper authorized access to the /prometheus endpoint. This change
allows defining a custom username/password pair in the configuration of
the plugin which can be used by the scraper to access only the
prometheus metrics and nothing else.